UNIX Pipeline-friendly asset update executable
Can now be run in a UNIX pipeline to update an asset from a line-based
input stream. SecurityCenter will optimize the IPs into ranges
automatically. You can add CIDRs, individual IPs, or ranges, and it will
figure out the rest.

See README.md for a fully worked example of this.